% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_detect.R
\name{detect_multi}
\alias{detect_multi}
\title{Detect multiple patterns using optimized flat array output}
\usage{
detect_multi(strings, patterns, output, engine, parallel, chunk_size)
}
\description{
Detect multiple patterns using optimized flat array output
}
