% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_detect.R
\name{detect_single}
\alias{detect_single}
\title{Detect single pattern match (uses cached compilation)}
\usage{
detect_single(strings, pattern, engine, parallel)
}
\description{
Detect single pattern match (uses cached compilation)
}
