% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_string_detect_fancy_cached}
\alias{r_string_detect_fancy_cached}
\title{Fancy-regex single pattern with caching}
\usage{
r_string_detect_fancy_cached(strings, pattern, parallel)
}
\description{
Fancy-regex single pattern with caching
}
