% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_string_detect_multi_fancy_optimized}
\alias{r_string_detect_multi_fancy_optimized}
\title{Multi-pattern fancy-regex with caching}
\usage{
r_string_detect_multi_fancy_optimized(
  strings,
  patterns,
  parallel_strategy,
  chunk_size
)
}
\description{
Multi-pattern fancy-regex with caching
}
