% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{r_string_detect_regex_cached}
\alias{r_string_detect_regex_cached}
\title{OPTIMIZED: Single pattern detection}
\usage{
r_string_detect_regex_cached(strings, pattern, parallel)
}
\description{
OPTIMIZED: Single pattern detection
}
